requirements.txt
akismet==24.11.0
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.1
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
defusedxml==0.7.1
Django==4.2.25
django-appconf==1.1.0
django-bootstrap4==22.3
django-bootstrap4-form==4.0.2
django-cleanup==9.0.0
django-cogwheels==0.3
django-compressor==3.1
django-cors-headers==3.11.0
django-countries==7.3.2
django-crontab==0.7.1
django-debug-toolbar==3.2
django-environ==0.8.1
django-extensions==3.1.5
django-filter==2.4.0
django-formtools==2.3
django-groups-manager==1.3.0
django-guardian==2.4.0
django-ipware==4.0.2
django-jet-reboot==1.3.10
django-js-asset==2.2.0
django-markdown-deux==1.0.5
django-measurement==3.2.4
django-model-utils==5.0.0
django-mptt==0.17.0
django-oauth-toolkit==3.0.0
django-phonenumber-field==6.1.0
django-polymorphic==3.1.0
django-ratelimit==4.1.0
django-redis==5.2.0
django-reversion==5.0.12
django-select2==7.10.1
django-storages==1.11.1
django-tenant-users==1.0.0
django-tenants==3.8.0
django-timezone-field==4.2.1
django-user-accounts==3.0.4
django-widget-tweaks==1.4.12
django_sendmail_backend==0.1.2
djangorestframework==3.15.1
djangorestframework-simplejwt[crypto]==5.3.1
drf-spectacular==0.28.0
easy-thumbnails==2.8.3
email-reply-parser==0.5.12
firebase-admin==6.6.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==20.0.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
kombu==5.5.3
lxml==5.4.0
Markdown==3.8
markdown2==2.5.3
measurement==3.2.2
mpmath==1.3.0
msgpack==1.1.0
oauthlib==3.2.2
phonenumbers==8.12.45
pillow==11.2.1
pinax-invitations==7.0.0
pinax-teams==2.0.0
prompt_toolkit==3.0.51
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
rcssmin==1.1.0
redis==6.1.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.0
rpds-py==0.27.0
rsa==4.9
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.7
sqlparse==0.5.3
sweetify==2.3.1
sympy==1.14.0
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whois==0.9.13
python-dotenv==1.0.1